requirements.txt
discord.py==1.5.0
PyNaCl==1.4.0
cffi==1.14.3
youtube-dl==2020.3.8
python-dotenv==0.14.0
asyncpg==0.21.0
psutil==5.7.2